Implement separate file app #10
