+++ practice/src/dashboard.html
@@ -117,7 +117,7 @@
   </form>
   <!-- END MODAL -->
 
-  <script type="module" src="./typescript/app.ts"></script>
+  <script type="module" src="./typescript/dashboard-app.ts"></script>
 </body>
 <script>
