+++ practice/src/typescript/helpers/type-user.ts
@@ -1,0 +1,8 @@
+export interface User {
+  id: string;
+  email: string;
+  password: string;
+  username: string;
+  confirmPassword: string;
+  role: string;
+}
