+++ practice/src/typescript/models/recipe-model.ts
@@ -1,18 +1,4 @@
-export interface Recipe {
-  name: string;
-  description: string;
-  ingredient: string[];
-  instruction: string;
-  nutrition: string;
-  creator: string;
-  category: string;
-  id: string;
-  role: string;
-  imageURL: string;
-  ratings: number;
-  createdAt: Date;
-}
-
+import { Recipe } from "../helpers/type-recipe";
 export default class RecipeModel {
   private recipes: Recipe[];
